feat(discoveryDetailsPlaceholders): Began refactoring renderers into separate files
