+++ packages/frontend/src/features/Study/StudyDetails/StudyGroup.tsx
@@ -2,7 +2,7 @@ import React, { useMemo } from 'react';
 import { StudyTabGroup } from '../types';
 import { Text } from '@mantine/core';
 import { JSONObject } from '@gen3/core';
-import { createFieldRendererElement } from './StudyItems';
+import { createFieldRendererElement } from './StudyItems/StudyItems';
 import { JSONPath } from 'jsonpath-plus';
 
 interface StudyTabGroupProps extends StudyTabGroup {
@@ -37,7 +37,7 @@ const StudyGroup = ({ data, header, fields }: StudyTabGroupProps) => {
   return (
     <div className=" flex flex-col">
       <div className="bg-accent-lightest w-full p-1 mb-2">
-        <Text color="primary" fw={700}>
+        <Text c="primary" fw={700}>
           {header}
         </Text>
       </div>
