+++ packages/frontend/src/features/Study/StudyDetails/StudyItems/RendererFactory.tsx
