+++ packages/frontend/src/features/Study/StudyDetails/StudyItems/Renderers/BlockTextField.tsx
@@ -1,0 +1,15 @@
+import React, { ReactElement } from 'react';
+import { discoveryFieldStyle, jsonValueToElement } from './utils';
+import { JSONValue } from '@gen3/core';
+
+/**
+ * Renders a block of text.
+ *
+ * @param {JSONValue} fieldValue - The value to render.
+ * @returns {ReactElement} The rendered block of text.
+ */
+const BlockTextField = (fieldValue: JSONValue): ReactElement => (
+  <div className={discoveryFieldStyle}>{jsonValueToElement(fieldValue)}</div>
+);
+
+export default BlockTextField;
