+++ packages/frontend/src/features/Study/StudyDetails/StudyItems/Renderers/LinkField.tsx
@@ -1,0 +1,12 @@
+import Link from 'next/link';
+import { Text } from '@mantine/core';
+
+const LinkField = (linkValue: string) => (
+  <Link href={linkValue} target="_blank" rel="noreferrer">
+    <Text c="utility.0" className="underline">
+      {linkValue}
+    </Text>
+  </Link>
+);
+
+export default LinkField;
